Refactor for easier deployment.
This moves the various cli code from the server directory to the
histomicstk/cli directory.  As such, when installed without girder, the
cli code is available.

Although MANIFEST.in has been altered, `python setup.py sdist` doesn't
quite produce a fully working pypi package.  While this package includes
all relevant filse, when pip installed elsewhere with manually installed
prerequisites, for reasons I don't yet understand, the included xml cli
specifications are not copied into the deployed python packages.
