+++ histomicstk/cli/BackgroundIntensity/BackgroundIntensity.xml
