+++ histomicstk/cli/ColorDeconvolution/ColorDeconvolution.py
@@ -1,6 +1,3 @@
-import os
-import sys
-
 import skimage.io
 
 from ctk_cli import CLIArgumentParser
@@ -8,11 +5,10 @@ import large_image
 
 import histomicstk.preprocessing.color_deconvolution as htk_cd
 
+from histomicstk.cli import utils
+
 import logging
 logging.basicConfig()
-
-sys.path.append(os.path.normpath(os.path.join(os.path.dirname(__file__), '..')))
-from cli_common import utils  # noqa
 
 
 def main(args):
