+++ histomicstk/cli/ColorDeconvolution/ColorDeconvolution.xml
