+++ histomicstk/cli/ComputeNucleiFeatures/ComputeNucleiFeatures.py
@@ -1,5 +1,4 @@
 import os
-import sys
 import json
 import time
 
@@ -18,11 +17,10 @@ import large_image
 
 from ctk_cli import CLIArgumentParser
 
+from histomicstk.cli import utils as cli_utils
+
 import logging
 logging.basicConfig(level=logging.CRITICAL)
-
-sys.path.append(os.path.normpath(os.path.join(os.path.dirname(__file__), '..')))
-from cli_common import utils as cli_utils  # noqa
 
 
 def compute_tile_nuclei_features(slide_path, tile_position, args, it_kwargs,
