+++ histomicstk/cli/NucleiClassification/NucleiClassification.py
@@ -1,5 +1,4 @@
 import os
-import sys
 import json
 import colorsys
 
@@ -10,11 +9,10 @@ import dask.dataframe as dd
 
 from ctk_cli import CLIArgumentParser
 
+from histomicstk.cli import utils as cli_utils
+
 import logging
 logging.basicConfig(level=logging.CRITICAL)
-
-sys.path.append(os.path.normpath(os.path.join(os.path.dirname(__file__), '..')))
-from cli_common import utils as cli_utils  # noqa
 
 
 def gen_distinct_rgb_colors(n, seed=None):
