+++ histomicstk/cli/docker-entrypoint.sh
