+++ histomicstk/cli/slicer_cli_list.json
