+++ plugin_tests/color_normalization_test.py
@@ -17,7 +17,6 @@
 #  limitations under the License.
 ###############################################################################
 
-import sys
 import collections
 import numpy as np
 import os
@@ -27,8 +26,7 @@ import unittest
 from histomicstk.preprocessing import color_conversion as htk_cvt
 from histomicstk.preprocessing import color_normalization as htk_cn
 
-sys.path.append(os.path.normpath(os.path.join(os.path.dirname(__file__), '../server')))
-from cli_common import utils as cli_utils  # noqa
+from histomicstk.cli import utils as cli_utils
 
 
 TEST_DATA_DIR = os.path.join(os.environ['GIRDER_TEST_DATA_PREFIX'],
